requirements.txt
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Deprecated==1.2.14
distro==1.9.0
Django==4.2.9
django-celery-beat==2.5.0
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
kombu==5.3.5
magic-filter==1.0.12
multidict==6.0.4
netaddr==1.0.0
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
yookassa==3.0.1
